% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoformat.R
\name{gapit_geno_format}
\alias{gapit_geno_format}
\title{生成GAPIT的geno文件格式}
\usage{
gapit_geno_format(geno = snp50k_Williams_82.a24, threshold = 0.05)
}
\arguments{
\item{geno}{数据为Soybase网站下载的snp50k_Williams_82.a24数据，数据列为CHROM POS ID ...}

\item{threshhold}{删除稀有变异预值，默认为0.05。计算时不包含H和缺失}
}
\value{
返回用于gapit 计算的标准数据，其中材料已按名称升序排列
}
\description{
生成GAPIT的geno文件格式
}
